Code refactorings for better file organization
