+++ src/main/java/npc/kassinimvp/security/service/UserDetailsImpl.java
@@ -10,7 +10,7 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
-import npc.kassinimvp.entity.Location;
+import npc.kassinimvp.entity.definitions.Location;
 
 @ToString
 public class UserDetailsImpl implements UserDetails {
